% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{Prepare Data for use}
\usage{
dataPrep(df, dd = NULL, varLabel = NULL, dir = getwd())
}
\arguments{
\item{df}{asdf}

\item{dd}{asdf}

\item{varLabel}{asdf}

\item{dir}{asdf}
}
\value{
sdofi
}
\description{
This function will take a \code{data.frame} as an argument, and optionally a data
dictionary, and compile everything that the program needs to work, or export
a csv to the working directory to be filled in and later input
}
\examples{
#asdf
}
