% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{test}
\alias{test}
\title{test dataset}
\format{
A data.frame with 400 rows and 42 variables.
\describe{
\item{usubjid}{subject identifier}
}
}
\usage{
test
}
\description{
This is the documentation for the test dataset.
}
\keyword{datasets}
