% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{testdd}
\alias{testdd}
\title{test Data Dictionary}
\format{
A data.frame with 39 rows and 2 variables.
\describe{
\item{usubjid}{subject identifier}
}
}
\usage{
testdd
}
\description{
This is the documentation for the test dat aset.
}
\keyword{datasets}
