% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type<-.data.frame}
\alias{type<-.data.frame}
\title{Add type attribute to data.frame}
\usage{
\method{type}{data.frame}(x, fill = NULL, ...) <- value
}
\arguments{
\item{x}{asd}

\item{fill}{asd}

\item{...}{asd}

\item{value}{asd}
}
\value{
asd
}
\description{
Add type attribute to data.frame
}
\examples{
#asds
}
