% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type<-.default}
\alias{type<-.default}
\title{Assign type attribute}
\usage{
\method{type}{default}(x, ...) <- value
}
\arguments{
\item{x}{asd}

\item{...}{asd}

\item{value}{asd}
}
\value{
asd
}
\description{
Assign type attribute
}
