% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type.data.frame}
\alias{type.data.frame}
\title{Get type attributes}
\usage{
\method{type}{data.frame}(x, ...)
}
\arguments{
\item{x}{asd}

\item{...}{asd}
}
\value{
asd
}
\description{
Get type attributes
}
\examples{
#asd
}
