% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type.default}
\alias{type.default}
\title{Get type attribute}
\usage{
\method{type}{default}(x, ...)
}
\arguments{
\item{x}{asd}

\item{...}{asd}
}
\value{
asd
}
\description{
Get type attribute
}
\examples{
#asd
}
